+++ HexLibrary/LinearMemoryManager.h
@@ -17,12 +17,12 @@ namespace HL
 				U* m_data = nullptr;
 				size_t m_total_count = 0;
 				size_t m_used_count = 0;
-				float m_increasing_factor = 2.0f;
+				size_t m_increasing_factor = 2;
 
 				void Resize(size_t Count, bool AllowGrowth = true) {
 					size_t to = 0;
 					if (AllowGrowth)
-						to = (size_t)((m_total_count + Count)*m_increasing_factor);
+						to = (m_total_count + Count)* m_increasing_factor;
 					else
 						to = m_total_count + Count;
 					void*block = Allocator->Alloc(sizeof(U)*to);
